Added fullscreen support
