+++ myManga-App/MainWindow.xaml
@@ -2,7 +2,7 @@
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:p="http://schemas.microsoft.com/netfx/2007/xaml/presentation"
-    xmlns:window="clr-namespace:myManga_App.Themes.Window"
+    xmlns:window="clr-namespace:myManga_App.Controls.Window"
     xmlns:ViewModels="clr-namespace:myManga_App.ViewModels"
     xmlns:MVVM="clr-namespace:myManga_App.Objects.MVVM"
     xmlns:AssemblyInformation="clr-namespace:myManga_App.Objects.About"
@@ -23,7 +23,7 @@
     <window:myMangaWindow.Resources>
         <ResourceDictionary>
             <ResourceDictionary.MergedDictionaries>
-                <ResourceDictionary Source="/myManga;component/Themes/Window/myMangaWindow.xaml"/>
+                <ResourceDictionary Source="/myManga;component/Controls/Window/myMangaWindow.xaml"/>
                 <ResourceDictionary Source="/myManga;component/Themes/CoreResourceDictionary.xaml"/>
                 <ResourceDictionary Source="/myManga;component/Resources/Localization/Dictionary_en-US.xaml"/>
                 <ResourceDictionary Source="/myManga;component/Resources/ViewModel-View-Link.xaml"/>
