+++ myManga-App/myManga-App.csproj
@@ -110,7 +110,7 @@
     <Compile Include="Objects\UserInterface\SiteExtentionInformationObject.cs" />
     <Compile Include="Objects\UserConfig\UserConfigurationObject.cs" />
     <Compile Include="Themes\Button\MetroButton.cs" />
-    <Compile Include="Themes\Window\myMangaWindow.cs" />
+    <Compile Include="Controls\Window\myMangaWindow.cs" />
     <Compile Include="ViewModels\BaseViewModel.cs" />
     <Compile Include="ViewModels\HomeViewModel.cs" />
     <Compile Include="ViewModels\MainViewModel.cs" />
@@ -215,7 +215,7 @@
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="Themes\Window\myMangaWindow.xaml">
+    <Page Include="Controls\Window\myMangaWindow.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
